Autenticação nas rotas e Middleware

+++ src/avaliacao/avaliacao.controller.ts
@@ -7,17 +7,28 @@ import {
   Param,
   Delete,
   ParseIntPipe,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { AvaliacaoService } from './avaliacao.service';
 import { CreateAvaliacaoDto } from './dto/create-avaliacao.dto';
 import { UpdateAvaliacaoDto } from './dto/update-avaliacao.dto';
+import { CurrentUser } from 'src/auth/decorators/CurrentUser.decorator';
+import { UserPayload } from 'src/auth/types/UserPayload';
 
 @Controller('avaliacao')
 export class AvaliacaoController {
   constructor(private readonly avaliacaoService: AvaliacaoService) {}
 
   @Post()
-   async create(@Body() createAvaliacaoDto: CreateAvaliacaoDto) {
+  async create(
+    @Body() createAvaliacaoDto: CreateAvaliacaoDto,
+    @CurrentUser() currentUser: UserPayload,
+  ) {
+    if (createAvaliacaoDto.usuarioID !== currentUser.sub) {
+      throw new UnauthorizedException(
+        'Avaliação não pode ser criada para outro usuário',
+      );
+    }
     return await this.avaliacaoService.create(createAvaliacaoDto);
   }
 
@@ -35,12 +46,29 @@ export class AvaliacaoController {
   async update(
     @Param('id', ParseIntPipe) id: number,
     @Body() updateAvaliacaoDto: UpdateAvaliacaoDto,
+    @CurrentUser() currentUser: UserPayload,
   ) {
+    const avaliacao = await this.avaliacaoService.findOne(id);
+    if (avaliacao.usuarioID !== currentUser.sub) {
+      throw new UnauthorizedException(
+        'Avaliação não pode ser criada para outro usuário',
+      );
+    }
     return this.avaliacaoService.update(id, updateAvaliacaoDto);
   }
 
   @Delete(':id')
-  async remove(@Param('id', ParseIntPipe) id: number) {
+  async remove(
+    @Param('id', ParseIntPipe) id: number,
+    @CurrentUser() currentUser: UserPayload,
+  ) {
+    const avaliacao = await this.avaliacaoService.findOne(id);
+    if (avaliacao.usuarioID !== currentUser.sub) {
+      throw new UnauthorizedException(
+        'Avaliação não pode ser criada para outro usuário',
+      );
+    }
+
     return await this.avaliacaoService.remove(id);
   }
 }
